% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAIBSARNA.R
\name{buildAIBSARNA}
\alias{buildAIBSARNA}
\title{Function to create an ExpressionSet containing BrainSpan Data}
\usage{
buildAIBSARNA(mini = FALSE)
}
\arguments{
\item{mini}{Default is FALSE.  If \code{mini=TRUE}, build a miniature 
version of AIBSARNA that does not require internet connectivity and is
suitable for example purposes only}
}
\value{
A Biobase ExpressionSet containing BrainSpan data, with the addition
    of RefSeq IDs via biomaRt
}
\description{
This function is used to build the AIBSARNA ExpressionSet, and must be run 
prior to running any other function in \pkg{BrainSABER}. This function 
will download the data from http://brainspan.org and may take several 
minutes, depending on internet connection speeds.
}
\examples{
AIBSARNA <- buildAIBSARNA(mini = TRUE)
}
