% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistance.R
\name{calcDistance}
\alias{calcDistance}
\title{Calculate Correlation Between User Dataset and Allen Brain Institute in Shiny App}
\usage{
calcDistance(dataSet)
}
\arguments{
\item{dataSet}{a CellScabbard object with non-empty relevantGenes slot}
}
\value{
a list of correlation data frames and matrices, one list element for each sample
}
\description{
This function takes a CellScabbard object as returned by
\code{mapIDs} and returns list of matrices/data frame for each sample. This list
includes the unordered dataframe of correlation values, significance, and metadata;
and one matrix each for Kendall's tau and Spearman's rho correlation coefficients across
brain structures and ages in the Allen Brain Developmental Transcriptome. This function
is to be used in the context of the Shiny application workflow.
}
