% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellScabbard-methods.R
\name{CellScabbard-methods}
\alias{CellScabbard-methods}
\alias{dataSetId}
\alias{CellScabbard,ANY,ANY-method}
\alias{dataSetId<-}
\alias{AIBSARNAid}
\alias{AIBSARNAid<-}
\alias{relevantGenes}
\alias{relevantGenes<-}
\alias{similarityScores}
\alias{similarityScores<-}
\alias{similarityDFs}
\alias{similarityDFs<-}
\alias{similarityMatrices}
\alias{similarityMatrices<-}
\alias{UNDmatrices}
\alias{UNDmatrices<-}
\title{Methods for the CellScabbard class}
\usage{
dataSetId(cs)

dataSetId(cs) <- value

AIBSARNAid(cs)

AIBSARNAid(cs) <- value

relevantGenes(cs)

relevantGenes(cs) <- value

similarityScores(cs)

similarityScores(cs) <- value

similarityDFs(cs)

similarityDFs(cs) <- value

similarityMatrices(cs)

similarityMatrices(cs) <- value

UNDmatrices(cs)

UNDmatrices(cs) <- value
}
\arguments{
\item{cs}{A CellScabbard object}

\item{value}{data type, any of matrix, data.frame,list, or SimpleList}
}
\value{
The contents of a slot of the CellScabbard object
}
\description{
These methods operate on CellScabbard objects.  They are used to access the
results of the \pkg{BrainSABER} workflow stored within a CellScabbard.
}
\examples{
# construct example data set
AIBSARNA <- buildAIBSARNA(mini = TRUE)

# get a random sample of 3 genes
totalGenes <- nrow(AIBSARNA)
gene_idx <- sample.int(totalGenes, 3)
sample_idx <- c(1,3,5)

# Subset AIBSARNA
exprs <- assay(AIBSARNA)[gene_idx, sample_idx]
fd <- rowData(AIBSARNA)[gene_idx, ]
pd <- colData(AIBSARNA)[sample_idx, ]

# construct a CellScabbard data set
myGenes <- CellScabbard(exprsData = exprs, phenoData = pd, featureData = fd, 
                        AIBSARNA = AIBSARNA, autoTrim = TRUE)
relevantGenes(myGenes)

# the following fields will be empty as output must be assigned to 
# them first
similarityScores(myGenes)
similarityMatrices(myGenes)
similarityDFs(myGenes)
UNDmatrices(myGenes)
}
