% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellScabbard-class.R
\docType{class}
\name{CellScabbard}
\alias{CellScabbard}
<<<<<<< HEAD
\alias{.CellScabbard}
\alias{CellScabbard-class}
\title{The CellScabbard class}
\usage{
CellScabbard(
  exprsData,
  phenoData = NULL,
  featureData = NULL,
  AIBSARNA = NULL,
  autoTrim = TRUE
)
}
\arguments{
\item{exprsData}{expression data matrix for an experiment}

\item{phenoData}{a data frame containing attributes of individual
cells}

\item{featureData}{a data frame containing attributes of features
(e.g. genes)}

\item{AIBSARNA}{an instance of the AIBSARNA dataset, built using the 
\code{buildAIBSARNA()} function}

\item{autoTrim}{If TRUE (default), automatically trim user data to match
AIBSARNA using the column names of AIBSARNA and featureData (or rownames
of exprsData, if featureData is not supplied) which produce the most 
matched identifiers.  Also automatically fills the relevantGenes slot,
using the same column names.  The column names are stored in the 
dataSetId and AIBSARNAid slots.}
}
\value{
a new CellScabbard object
}
\description{
The main class used by the \pkg{BrainSABER} package to hold single cell 
expression data, relevant gene data, and similarity matrices.
CellScabbard extends the 
[SummarizedExperiment::SummarizedExperiment-class] class.
=======
\alias{CellScabbard-class}
\title{The CellScabbard class}
\description{
The main class used by the \pkg{BrainSABER} package to hold single cell 
expression data, relevant gene data, and similarity matrices.
CellScabbard extends the basic [Biobase::ExpressionSet-class] class.
>>>>>>> refs/remotes/origin/master
}
\details{
This class is initialized from a matrix of gene expression values and
associated metadata.  Methods that operate on CellScabbard objects
comprise the BrainSABER workflow.
}
\section{Fields}{

\describe{
<<<<<<< HEAD
\item{\code{dataSetId}}{A character vector of length 1, containing the column name
of the user's data used to trim the data to match AIBSARNA.}

\item{\code{AIBSARNAid}}{A character vector of length 1, containing the column name
of AIBSARNA used to trim AIBSARNA to match the user's data}

\item{\code{relevantGenes}}{A [SummarizedExperiment::SummarizedExperiment-class] 
containing a subset of data from AIBSARNA. Generated by the 
\pkg{BrainSABER} workflow.}
=======
\item{\code{relevantGenes}}{A [Biobase::ExpressionSet-class] containing a
subset of data from \pkg{AIBSARNA}. Generated by the \pkg{BrainSABER}
workflow.}
>>>>>>> refs/remotes/origin/master

\item{\code{similarityScores}}{A \code{data.frame} containing similarity
scores. Generated by the \pkg{BrainSABER} workflow.}

<<<<<<< HEAD
\item{\code{similarityDFs}}{A \code{list} containing similarity data frames with
age, structure, and similarity scores sorted in decreasing order.
Generated by the \pkg{BrainSABER} workflow.}

\item{\code{similarityMatrices}}{A [SummarizedExperiment::Assays-class] object 
containing similarity matrices, with identical dimensions, for each 
sample in phenoData. Generated by the \pkg{BrainSABER} workflow.}

\item{\code{UNDmatrices}}{A [SummarizedExperiment::Assays-class] object containing
identical-dimension UND matrices for each sample in phenoData.
Generated by the \pkg{BrainSABER} workflow.}
}}

\examples{
# construct example data set
AIBSARNA <- buildAIBSARNA(mini = TRUE)

# get a random sample of 3 genes
totalGenes <- nrow(AIBSARNA)
gene_idx <- sample.int(totalGenes, 3)
sample_idx <- c(1,3,5)

# Subset AIBSARNA
exprs <- assay(AIBSARNA)[gene_idx, sample_idx]
fd <- rowData(AIBSARNA)[gene_idx, ]
pd <- colData(AIBSARNA)[sample_idx, ]

# construct a CellScabbard data set
myGenes <- CellScabbard(exprsData = exprs, phenoData = pd, featureData = fd, 
                        AIBSARNA = AIBSARNA, autoTrim = TRUE)
}
=======
\item{\code{similarityMatrices}}{A [Biobase::AssayData-class] object containing
similarity matrices, with identical dimensions, one for each sample in
phenoData. Generated by the \pkg{BrainSABER} workflow.}

\item{\code{UNDmatrices}}{A [Biobase::AssayData-class] object containing
identical-dimension UND matrices, one for each sample in phenoData.
Generated by the \pkg{BrainSABER} workflow.}
}}

>>>>>>> refs/remotes/origin/master
