% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getASCdf.R
\name{getASCdf}
\alias{getASCdf}
\title{Get Age, Structure Acronym, and Cosine Similarity Data Frame}
\usage{
getASCdf(relevantGenes)
}
\arguments{
\item{relevantGenes}{a Biobase Expression set}
}
\value{
a three-column data.frame
}
\description{
This function takes a \code{Biobase ExpressionSet} as returned by
\code{getCosineSim} and constructs a \code{data.frame} with columns age,
structure_acroynym, and cosine_similarity, sorted by cosine_similarity in
decreasing order.
}
