% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getASCmat.R
\name{getASCmat}
\alias{getASCmat}
\title{Get Age, Structure Acronym, and Cosine Similarity Matrix}
\usage{
getASCmat(relevantGenes)
}
\arguments{
\item{relevantGenes}{a Biobase ExpressionSet}
}
\value{
a numeric matrix of cosine similaritys
}
\description{
This function takes a \code{Biobase ExpressionSet} as returned by
\code{getCosineSim} and returns a numeric matrix of cosine_similarity with
rows labeled by age, and columns labeled by structure_acronym.
}
