% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCosineSim.R
\name{getCosineSim}
\alias{getCosineSim}
\title{Get Cosine Similarity Scoring for a Gene Expression Vector}
\usage{
getCosineSim(v, dataSet)
}
\arguments{
\item{v}{a vector of HGNC-conformant named gene expression values}

\item{dataSet}{the AIBSARNA dataSet}
}
\value{
a Biobase ExpressionSet, with similarity scores stored as part of the
   sample data (phenoData)
}
\description{
This function computes the cosine similarity score of a gene expression
vector as compared to the AIBSARNA dataset, though other dataSets may be
supported in the future.  The vector must consist of numerical expression
values named with HGNC-conformant gene names.
}
