% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimDataFrame.R
\name{getSimDataFrame}
\alias{getSimDataFrame}
\title{Get Age, Structure Acronym, and Similarity Scores Data Frame}
\usage{
getSimDataFrame(relevantGenes, similarity_method = "cosine")
}
\arguments{
\item{relevantGenes}{a Biobase Expression set}

\item{similarity_method}{currently supported similarity methods are "cosine"
and "euclidean", defaults to "cosine"}
}
\value{
a three-column data.frame
}
\description{
This function takes a \code{Biobase ExpressionSet} as returned by
\code{getSimScores} and constructs a \code{data.frame} with columns age,
structure_acroynym, and either cosine_similarity or euclidean_similarity,
sorted by similarity score in decreasing order.
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes)
myGeneSet <- getSimScores(myGenes, myGeneSet, similarity_method = "cosine")
myGeneSet <- getSimScores(myGenes, myGeneSet, similarity_method = "euclidean")
myCosineDF <- getSimDataFrame(myGeneSet, similarity_method = "cosine")
myEuclideanDF <- getSimDataFrame(myGeneSet, similarity_method = "euclidean")
}
