% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newCellScabbard.R
\name{newCellScabbard}
\alias{newCellScabbard}
\title{Creates a new CellScabbard object.}
\usage{
newCellScabbard(exprsData, phenoData = NULL, featureData = NULL)
}
\arguments{
\item{exprsData}{expression data matrix for an experiment}

\item{phenoData}{an AnnotatedDataFrame containing attributes of individual
cells}

\item{featureData}{an AnnotatedDataFrame containing attributes of features
(e.g. genes)}
}
\value{
a new CellScabbard object
}
\description{
Creates a new CellScabbard object.
}
