% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBrainSABER.R
\name{runShinyBrainSABER}
\alias{runShinyBrainSABER}
\title{ShinyBrainSABER}
\usage{
runShinyBrainSABER()
}
\value{
NULL (Invisibly)
}
\description{
This function runs the Shiny app for the BrainSABER workflow.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
    options(device.ask.default = FALSE)
    runShinyBrainSABER()
}
}
