% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppServer.R
\name{shinyAppServer}
\alias{shinyAppServer}
\title{server side logic for the shiny app}
\usage{
shinyAppServer(input, output)
}
\arguments{
\item{input}{shiny server input, provided automatically}

\item{output}{shiny server output, provided automatically}
}
\value{
a shiny server object
}
\description{
server side logic for the shiny app
}
