% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppUI.R
\docType{data}
\name{sidebar}
\alias{sidebar}
\title{Html blocks for building shiny app User Interface}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
sidebar
}
\description{
Html blocks for building shiny app User Interface
}
\keyword{datasets}
