% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellScabbard-class.R
\docType{class}
\name{CellScabbard}
\alias{CellScabbard}
\alias{CellScabbard-class}
\title{The CellScabbard class}
\description{
The main class used by the \pkg{BrainSABER} package to hold single cell 
expression data, relevant gene data, and similarity matrices.
CellScabbard extends the basic [Biobase::ExpressionSet-class] class.
}
\details{
This class is initialized from a matrix of gene expression values and
associated metadata.  Methods that operate on CellScabbard objects
comprise the BrainSABER workflow.
}
\section{Fields}{

\describe{
\item{\code{relevantGenes}}{A [Biobase::ExpressionSet-class] containing a
subset of data from \pkg{AIBSARNA}. Generated by the \pkg{BrainSABER}
workflow.}

\item{\code{similarityScores}}{A \code{data.frame} containing similarity
scores. Generated by the \pkg{BrainSABER} workflow.}

\item{\code{similarityMatrices}}{A [Biobase::AssayData-class] object containing
similarity matrices, with identical dimensions, one for each sample in
phenoData. Generated by the \pkg{BrainSABER} workflow.}

\item{\code{UNDmatrices}}{A [Biobase::AssayData-class] object containing
identical-dimension UND matrices, one for each sample in phenoData.
Generated by the \pkg{BrainSABER} workflow.}
}}

