% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalVector.R
\name{getExternalVector}
\alias{getExternalVector}
\title{getExternalVector}
\usage{
getExternalVector(dataSet, index = 1, dataSetId = "gene_symbol",
  AIBSARNAid = c("gene_id", "ensembl_gene_id", "gene_symbol", "entrez_id",
  "refseq_ids"))
}
\arguments{
\item{dataSet}{a Biobase ExpressionSet}

\item{index}{the integer index of the sample of dataSet to be used}

\item{dataSetId}{the name of the column of gene identifiers in fData(dataSet)
to be used to compare dataSet to AIBSARNA.}

\item{AIBSARNAid}{the name of the column of fData(AIBSARNA) that is
comparable to dataSetId.  One of "gene_id", "ensembl_gene_id",
"gene_symbol", "entrez_id", "refseq_ids"}
}
\value{
a named vector of gene expression values
}
\description{
Get a named vector of gene expression values from a single sample in an
outside Biobase ExpressionSet, for use in creating subsets of AIBSARNA with
 \code{getRelevantGenes} and comparison with that subset with
 \code{getSimScores}
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes, gene_names = "HGNC")
myGeneSampleVector <- getExternalVector(myGeneSet, index = 1,
     dataSetId = "gene_symbol", AIBSARNAid = "gene_symbol")

}
