% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelevantGenes.R
\name{getRelevantGenes}
\alias{getRelevantGenes}
\title{Get a subset of AIBSARNA using a Gene Expression Vector}
\usage{
getRelevantGenes(v)
}
\arguments{
\item{v}{a vector of HGNC-conformant named gene expression values}
}
\value{
a Biobase ExpressionSet consisting of genes in v
}
\description{
This function returns a subset of the AIBSARNA dataset, containing only
the genes listed in the parameter vector, v.  The vector must consist of
numerical gene expression values named with HGNC-conformant gene names.
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes)

}
