% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimDataFrame.R
\name{getSimDataFrame}
\alias{getSimDataFrame}
\title{Get Age, Structure Acronym, and Similarity Scores Data Frame}
\usage{
getSimDataFrame(sim_score, relevantGenes, similarity_method = "cosine")
}
\arguments{
\item{sim_score}{a vector or data frame of similarity scores}

\item{relevantGenes}{a Biobase Expression set created using the 
\code{getRelevantGenes()} function}

\item{similarity_method}{currently supported similarity methods are "cosine"
and "euclidean", defaults to "cosine"}
}
\value{
a three-column data.frame or list of data frames
}
\description{
This function takes a similarity vector or data frame as returned by
\code{getSimScores} and a subset of AIBSARNA as returned by
\code{getRelevantGenes} and constructs a \code{data.frame} with columns age,
structure_acroynym, and either cosine_similarity or euclidean_similarity,
sorted by similarity score in decreasing order.  In the case of a similarity
data frame, a list of data frames is returned.
}
\examples{
AIBSARNA <- buildAIBSARNA(mini = TRUE)
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TSPAN6", "DPM1", "C1orf112")
myGeneSet <- getRelevantGenes(myGenes, AIBSARNA = AIBSARNA,
    AIBSARNAid = "gene_symbol")
myCosScore <- getSimScores(myGenes, myGeneSet, similarity_method = "cosine")
myEucScore <- getSimScores(myGenes, myGeneSet, 
    similarity_method = "euclidean")
myCosineDF <- getSimDataFrame(myCosScore, myGeneSet,
    similarity_method = "cosine")
myEuclideanDF <- getSimDataFrame(myEucScore, myGeneSet,
    similarity_method = "euclidean")
}
