% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimMatrix.R
\name{getSimMatrix}
\alias{getSimMatrix}
\title{Get Age, Structure Acronym, and Similarity Score Matrix}
\usage{
getSimMatrix(sim_score, relevantGenes)
}
\arguments{
\item{sim_score}{a vector of similarity scores}

\item{relevantGenes}{a Biobase ExpressionSet}
}
\value{
a numeric matrix of similarity scores
}
\description{
This function takes a similarity vector as returned by \code{getSimScores}
and returns a numeric matrix of similarity scores with rows labeled by age,
and columns labeled by structure_acronym.
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes)
myCosScore <- getSimScores(myGenes, myGeneSet, similarity_method = "cosine")
myEucScore <- getSimScores(myGenes, myGeneSet, similarity_method = "euclidean")
myCosineMatrix <- getSimMatrix(myCosScore, myGeneSet)
myEuclideanMatrix <- getSimMatrix(myEucScore, myGeneSet)
}
