% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimMatrix.R
\name{getSimMatrix}
\alias{getSimMatrix}
\title{Get Age, Structure Acronym, and Similarity Score Matrix}
\usage{
getSimMatrix(relevantGenes, similarity_method = "cosine")
}
\arguments{
\item{relevantGenes}{a Biobase ExpressionSet}

\item{similarity_method}{currently supported similarity methods are "cosine"
and "euclidean", defaults to "cosine"}
}
\value{
a numeric matrix of similarity scores
}
\description{
This function takes a \code{Biobase ExpressionSet} as returned by
\code{getSimScores} and returns a numeric matrix of similarity scores with
rows labeled by age, and columns labeled by structure_acronym.
}
\examples{
\dontrun{myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes)
myGeneSet <- getSimScores(myGenes, myGeneSet, similarity_method = "cosine")
myGeneSet <- getSimScores(myGenes, myGeneSet, similarity_method = "euclidean")
myCosineMatrix <- getSimMatrix(myGeneSet, similarity_method = "cosine")
myEuclideanMatrix <- getSimMatrix(myGeneSet, similarity_method = "euclidean")}
}
