% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimScores.R
\name{getSimScores}
\alias{getSimScores}
\title{Get Similarity Scoring for a Gene Expression Vector}
\usage{
getSimScores(data, relevantGenes, similarity_method = "cosine")
}
\arguments{
\item{data}{a named vector of gene expression values returned by
\code{getExternalVector}, or an ExpressionSet returned by
\code{getTrimmedExternalSet}}

\item{relevantGenes}{an ExpressionSet returned by \code{getRelevantGenes}}

\item{similarity_method}{currently supported similarity methods are "cosine"
and "euclidean", defaults to "cosine"}
}
\value{
If \code{data} is a vector, returns a vector of similarity scores for
    each sample in relevantGenes. If \code{data} is an ExpressionSet, returns
    a data frame, with columns containing the similarity scores for and named
    after each sample in \code{data}, and rows named after each sample in
    \code{relevantGenes}
}
\description{
This function computes the similarity score of a gene expression
vector returned by \code{getExternalVector} or a trimmed data set returned by
\code{getTrimmedExternalSet}, compared to a subset of AIBSARNA,
obtained by \code{getRelevantGenes}.
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes, AIBSARNAid = "gene_symbol")
CosScores <- getSimScores(myGenes, myGeneSet, similarity_method = "cosine")
EucScores <- getSimScores(myGenes, myGeneSet, similarity_method = "euclidean")
myCosSimFrame <- getSimScores(myGeneSet, myGeneSet)
}
