% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimScores.R
\name{getSimScores}
\alias{getSimScores}
\title{Get Similarity Scoring for a Gene Expression Vector}
\usage{
getSimScores(v, relevantGenes, similarity_method = "cosine")
}
\arguments{
\item{v}{a vector of HGNC-conformant named gene expression values}

\item{relevantGenes}{an ExpressionSet returned by \code{getRelevantGenes}}

\item{similarity_method}{currently supported similarity methods are "cosine"
and "euclidean", defaults to "cosine"}
}
\value{
relevantGenes, with similarity scores stored as part of the sample
    data (phenoData)
}
\description{
This function computes the similarity score of a gene expression
vector as compared to an ExpressionSet of genes, obtained by
\code{getRelevantGenes}.  The vector must consist of numerical expression
values named with HGNC-conformant gene names, and should be the same vector
used to generate the ExpressionSet.
}
\examples{
\dontrun{myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes)
myGeneSet <- getSimScores(myGenes, myGeneSet, similarity_method = "cosine")
myGeneSet <- getSimScores(myGenes, myGeneSet, similarity_method = "euclidean")}
}
