% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrimmedExternalSet.R
\name{getTrimmedExternalSet}
\alias{getTrimmedExternalSet}
\title{Get a trimmed version of an external data set}
\usage{
getTrimmedExternalSet(dataSet, dataSetId = "gene_symbol",
  AIBSARNAid = c("gene_id", "ensembl_gene_id", "gene_symbol",
  "entrez_id", "refseq_ids"))
}
\arguments{
\item{dataSet}{a Biobase ExpressionSet compatible data set}

\item{dataSetId}{the name of the column of gene identifiers in fData(dataSet)
to be used to compare dataSet to AIBSARNA.}

\item{AIBSARNAid}{the name of the column of fData(AIBSARNA) that is
comparable to dataSetId.  One of "gene_id", "ensembl_gene_id",
"gene_symbol", "entrez_id", "refseq_ids"}
}
\value{
a Biobase ExpressionSet
}
\description{
Returns an ExpressionSet that is a subset of \code{dataSet} containing only
genes that are present in AIBSARNA, for use in \code{getSimScores} or
\code{getUNDmatrix}.
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes, AIBSARNAid = "gene_symbol")
myTrimmedGeneSet <- getTrimmedExternalSet(myGeneSet,
     dataSetId = "gene_symbol", AIBSARNAid = "gene_symbol")
}
