% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUNDmatrix.R
\name{getUNDmatrix}
\alias{getUNDmatrix}
\title{getUNDmatrix}
\usage{
getUNDmatrix(dataSet, method = "discrete", up_threshold, down_threshold,
  matrix_type = c("num", "char"))
}
\arguments{
\item{dataSet}{a Biobase ExpressionSet}

\item{method}{currently only "discrete" is available}

\item{up_threshold}{a numerical value defining the lower bound (inclusive) by
which to consider a gene up-regulated}

\item{down_threshold}{a numerical value defining the lower bound (inclusive)
by which to consider a gene down-regulated}

\item{matrix_type}{either "num" for a numerical matrix with -1 indicating
down-regulation, 1 indicating up-regulation, and 0 indicating normal, or
"char" for a character matrix with "D" indicating down-regulation, "U"
indicating up-regulation, and "N" indicating normal}
}
\value{
either a numerical or character matrix
}
\description{
This function returns a matrix showing whether gene expression values in
 \code{dataSet} are up-regulated, down-regulated, or normal. \code{method = "discrete"} requires \code{up_threshold} and \code{down_threshold} to be defined.
}
\examples{
myGenes <- c(4.484885, 0.121902, 0.510035)
names(myGenes) <- c("TNFRSF1A", "BCL3", "NEFH")
myGeneSet <- getRelevantGenes(myGenes, gene_names = "HGNC")
myUNDnumericalMatrix <- getUNDmatrix(myGeneSet, method = "discrete",
    up_threshold = 3, down_threshold = 1, matrix_type = "num")
myUNDcharacterMatrix <- getUNDmatrix(myGeneSet, method = "discrete",
    up_threshold = 3, down_threshold = 1, matrix_type = "char")
}
