% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{mapIDs}
\alias{mapIDs}
\title{Form CellScabbard object and identify relevant gene subset in Shiny App}
\usage{
mapIDs(dataSet, AIBSARNA)
}
\arguments{
\item{dataSet}{a dataframe containing ENSEMBL gene ids and a gene expression matrix}

\item{AIBSARNA}{an instance of the AIBSARNA dataset}
}
\value{
a CellScabbard object containing a subset of relevant genes from the
Allen institute data
}
\description{
This function takes a dataframe read in from a csv file with the first column
representing ENSEMBL gene id and subsequent columns representing a normalized
gene expression matrix with one column for each sample and one row for each gene.
}
