% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBrainSABER.R
\name{runShinyBrainSABER}
\alias{runShinyBrainSABER}
\title{ShinyBrainSABER}
\usage{
runShinyBrainSABER()
}
\description{
This function runs the Shiny app for the BrainSABER workflow.
}
\examples{
\dontrun{
  library(BrainSABER)
  runShinyBrainSABER()
}
}
