% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBrainSABER.R
\name{runShinyBrainSaber}
\alias{runShinyBrainSaber}
\title{Run the shiny application for the BrainSABER package workflow.}
\usage{
runShinyBrainSaber()
}
\value{
NULL (Invisibly)
}
\description{
Run the shiny application for the BrainSABER package workflow.
}
\examples{
if (interactive()) {
    options(device.ask.default = FALSE)
    runShinyBrainSABER()
}
}
